requirements.txt
# FastAPI
uvicorn[standard]>=0.29.0
fastapi>=0.110.0
python-multipart>=0.0.9
jinja2>=3.1.3
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# PubMed API (using urllib instead of biopython for better compatibility)
xmltodict>=0.13.0

# LLM
openai>=1.30.0
httpx>=0.27.0

# Scheduling
apscheduler>=3.10.4

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
croniter>=2.0.1

# Data Processing
python-dateutil>=2.8.2
markdown>=3.5.2
beautifulsoup4>=4.12.3
